Refactor code, style, add features
I reworked the code to make it easier to work with. While I was doing
so, I added the ability to backspace, use periods to designate decimals,
as well as letting the user use the keyboard to control the calculator.
I also stylized the display to make it look nice. Since the display can
expand to show bigger numbers, I offset the calculator from the bottom,
in such a way that the buttons don't change position as the display gets
bigger.